sw: add nv user data area and merge firmware update command with nv write
This commit adds support for a 32-byte user-defined NV area. Write access
to the NV area is authenticated through a SHA-256 hash (to write to the area
one has to show the hash pre-image).

Additionally the firmware update command is merged with the NV write command
(firmware update mode can now be entered by writing to NV index 0xFA).